Make post_xorg_start hook part of pre_xorg_start
Because the pre_xorg_start hook will now be called from
/etc/X11/xinit/xinitrc.d. This means that the pre_xorg_start hook will
now be executed between session startup and X server startup stages. And
all commands executed in it will be executed in user mode, so we have to
rely on daemon.py to do the saving of the configuration files.